refactor(worktree-io): own worker handler and sessions
